requirements.txt
annotated-types==0.6.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyparsing==3.1.2
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1